oumi/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
oumi>=0.5.0

tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-c>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-c-sharp>=0.20.0
tree-sitter-php>=0.20.0
tree-sitter-ruby>=0.20.0
tree-sitter-bash>=0.20.0
tree-sitter-html>=0.20.0
tree-sitter-css>=0.20.0
tree-sitter-dockerfile>=0.2.0
tree-sitter-kotlin>=0.3.0
tree-sitter-json>=0.20.0
tree-sitter-sql>=0.3.0
